[refactor] todo 컴포넌트 분리
