+++ eslint.config.js
@@ -14,6 +14,7 @@ export default [
     rules: {
       "react/react-in-jsx-scope": "off",
       "react/jsx-no-target-blank": "off",
+      "react/prop-types": "off",
     }
   }
 ];
